# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: IRUSA Smart Event & Donor Intelligence Dashboard API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /events/{eventId}/logistics:
    get:
      operationId: listLogisticsTasks
      tags: [logistics]
      summary: List logistics tasks for an event
      parameters:
        - in: path
          name: eventId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of logistics tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LogisticsTask"
    post:
      operationId: createLogisticsTask
      tags: [logistics]
      summary: Create a logistics task
      parameters:
        - in: path
          name: eventId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLogisticsTaskBody"
      responses:
        "201":
          description: Created task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LogisticsTask"
  /logistics/{id}:
    patch:
      operationId: updateLogisticsTask
      tags: [logistics]
      summary: Update a logistics task
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLogisticsTaskBody"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LogisticsTask"
    delete:
      operationId: deleteLogisticsTask
      tags: [logistics]
      summary: Delete a logistics task
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ==================== REVENUE ====================
  /events/{eventId}/revenue:
    get:
      operationId: listRevenueEntries
      tags: [revenue]
      summary: List revenue entries for an event
      parameters:
        - in: path
          name: eventId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Revenue entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RevenueEntry"
    post:
      operationId: createRevenueEntry
      tags: [revenue]
      summary: Create revenue entry
      parameters:
        - in: path
          name: eventId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRevenueEntryBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RevenueEntry"
  /revenue/{id}:
    patch:
      operationId: updateRevenueEntry
      tags: [revenue]
      summary: Update revenue entry
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRevenueEntryBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RevenueEntry"
    delete:
      operationId: deleteRevenueEntry
      tags: [revenue]
      summary: Delete revenue entry
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ==================== FRF ====================
  /events/{eventId}/frf:
    get:
      operationId: getFRFRecord
      tags: [frf]
      summary: Get FRF record for an event
      parameters:
        - in: path
          name: eventId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: FRF record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FRFRecord"
        "404":
          description: Not found
    post:
      operationId: createFRFRecord
      tags: [frf]
      summary: Create or update FRF record
      parameters:
        - in: path
          name: eventId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFRFBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FRFRecord"
  /frf/{id}:
    patch:
      operationId: updateFRFRecord
      tags: [frf]
      summary: Update FRF record
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateFRFBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FRFRecord"

  # ==================== ALLOCATIONS ====================
...
components:
  schemas:
...
    LogisticsTask:
      type: object
      properties:
        id:
          type: integer
        eventId:
          type: integer
        taskName:
          type: string
        assignedTo:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
        status:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - eventId
        - taskName
        - status
        - createdAt
    CreateLogisticsTaskBody:
      type: object
      properties:
        taskName:
          type: string
        assignedTo:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
        status:
          type: string
          default: "pending"
        notes:
          type: ["string", "null"]
      required:
        - taskName
    UpdateLogisticsTaskBody:
      type: object
      properties:
        taskName:
          type: string
        assignedTo:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
        status:
          type: string
        notes:
          type: ["string", "null"]

    # ==================== REVENUE ====================
    RevenueEntry:
      type: object
      properties:
        id:
          type: integer
        eventId:
          type: integer
        paymentType:
          type: string
        amount:
          type: number
        quantity:
          type: ["integer", "null"]
        receivedDate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        enteredBy:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - eventId
        - paymentType
        - amount
        - createdAt
    CreateRevenueEntryBody:
      type: object
      properties:
        paymentType:
          type: string
        amount:
          type: number
        quantity:
          type: ["integer", "null"]
        receivedDate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        enteredBy:
          type: ["string", "null"]
      required:
        - paymentType
        - amount
    UpdateRevenueEntryBody:
      type: object
      properties:
        paymentType:
          type: string
        amount:
          type: number
        quantity:
          type: ["integer", "null"]
        receivedDate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        enteredBy:
          type: ["string", "null"]

    # ==================== FRF ====================
    FRFRecord:
      type: object
      properties:
        id:
          type: integer
        eventId:
          type: integer
        cashAmount:
          type: number
        checkAmount:
          type: number
        mailedCheckAmount:
          type: number
        mailedReceiptAmount:
          type: number
        onlineAmount:
          type: number
        otherAmount:
          type: number
        totalAmount:
          type: number
        submittedBy:
          type: ["string", "null"]
        submittedDate:
          type: ["string", "null"]
        reconciliationStatus:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - eventId
        - cashAmount
        - checkAmount
        - mailedCheckAmount
        - mailedReceiptAmount
        - onlineAmount
        - otherAmount
        - totalAmount
        - reconciliationStatus
        - createdAt
    CreateFRFBody:
      type: object
      properties:
        cashAmount:
          type: number
          default: 0
        checkAmount:
          type: number
          default: 0
        mailedCheckAmount:
          type: number
          default: 0
        mailedReceiptAmount:
          type: number
          default: 0
        onlineAmount:
          type: number
          default: 0
        otherAmount:
          type: number
          default: 0
        submittedBy:
          type: ["string", "null"]
        submittedDate:
          type: ["string", "null"]
        reconciliationStatus:
          type: string
          default: "pending"
        notes:
          type: ["string", "null"]
    UpdateFRFBody:
      type: object
      properties:
        cashAmount:
          type: number
        checkAmount:
          type: number
        mailedCheckAmount:
          type: number
        mailedReceiptAmount:
          type: number
        onlineAmount:
          type: number
        otherAmount:
          type: number
        submittedBy:
          type: ["string", "null"]
        submittedDate:
          type: ["string", "null"]
        reconciliationStatus:
          type: string
        notes:
          type: ["string", "null"]

    # ==================== ALLOCATION ====================
...